+++ frontend/src/components/Header/Header.tsx
@@ -43,8 +43,8 @@ import SvgXIcon from '@icons/XIcon'
 import { useBillingHook } from '@pages/Billing/Billing'
 import { getTrialEndDateMessage } from '@pages/Billing/utils/utils'
 import useLocalStorage from '@rehooks/local-storage'
-import { useApplicationContext } from '@routers/OrgRouter/context/ApplicationContext'
-import { useGlobalContext } from '@routers/OrgRouter/context/GlobalContext'
+import { useApplicationContext } from '@routers/ProjectRouter/context/ApplicationContext'
+import { useGlobalContext } from '@routers/ProjectRouter/context/GlobalContext'
 import analytics from '@util/analytics'
 import { auth } from '@util/auth'
 import { isProjectWithinTrial } from '@util/billing/billing'
