+++ frontend/src/components/KeyboardShortcutsEducation/KeyboardShortcutsEducation.tsx
@@ -5,7 +5,7 @@ import TextHighlighter from '@components/TextHighlighter/TextHighlighter'
 import { Badge, Box } from '@highlight-run/ui'
 import SvgSearchIcon from '@icons/SearchIcon'
 import { PLAYER_SKIP_DURATION } from '@pages/Player/utils/PlayerHooks'
-import { useGlobalContext } from '@routers/OrgRouter/context/GlobalContext'
+import { useGlobalContext } from '@routers/ProjectRouter/context/GlobalContext'
 import analytics from '@util/analytics'
 import clsx from 'clsx'
 import { AnimatePresence, motion } from 'framer-motion'
