+++ frontend/src/pages/Alerts/AlertConfigurationCard/AlertConfigurationCard.tsx
@@ -17,7 +17,7 @@ import { namedOperations } from '@graph/operations'
 import { DiscordChannel, SessionAlertType } from '@graph/schemas'
 import { DiscordChannnelsSection } from '@pages/Alerts/AlertConfigurationCard/DiscordChannelsSection'
 import SyncWithSlackButton from '@pages/Alerts/AlertConfigurationCard/SyncWithSlackButton'
-import { useApplicationContext } from '@routers/OrgRouter/context/ApplicationContext'
+import { useApplicationContext } from '@routers/ProjectRouter/context/ApplicationContext'
 import { useParams } from '@util/react-router/useParams'
 import { Divider, Form, message } from 'antd'
 import clsx from 'clsx'
