+++ frontend/src/pages/Auth/AdminForm.tsx
@@ -54,6 +54,7 @@ export const AdminForm: React.FC = () => {
 	const [updateAdminAboutYouDetails] = useUpdateAdminAboutYouDetailsMutation()
 
 	if (admin?.about_you_details_filled) {
+		debugger
 		navigate('/setup')
 	}
