+++ frontend/src/pages/EmailOptOut/EmailOptOut.tsx
@@ -12,8 +12,8 @@ import {
 import { namedOperations } from '@graph/operations'
 import { EmailOptOutCategory } from '@graph/schemas'
 import { Box } from '@highlight-run/ui'
-import { ApplicationContextProvider } from '@routers/OrgRouter/context/ApplicationContext'
-import { GlobalContextProvider } from '@routers/OrgRouter/context/GlobalContext'
+import { ApplicationContextProvider } from '@routers/ProjectRouter/context/ApplicationContext'
+import { GlobalContextProvider } from '@routers/ProjectRouter/context/GlobalContext'
 import { message } from 'antd'
 import { useDialogState } from 'ariakit/dialog'
 import { H } from 'highlight.run'
