+++ frontend/src/pages/IntegrationsPage/components/common/useIntegration.ts
@@ -5,7 +5,7 @@ import {
 } from '@graph/hooks'
 import { namedOperations } from '@graph/operations'
 import { Exact, IntegrationType } from '@graph/schemas'
-import { useApplicationContext } from '@routers/OrgRouter/context/ApplicationContext'
+import { useApplicationContext } from '@routers/ProjectRouter/context/ApplicationContext'
 import { useCallback } from 'react'
 
 type SettingsLoadingState = {
