+++ frontend/src/routers/ProjectRouter/ProjectRouter.tsx
@@ -20,9 +20,9 @@ import { HighlightEvent } from '@pages/Player/HighlightEvent'
 import { NetworkResource } from '@pages/Player/Toolbar/DevToolsWindowV2/utils'
 import { usePlayerFullscreen } from '@pages/Player/utils/PlayerHooks'
 import useLocalStorage from '@rehooks/local-storage'
-import { GlobalContextProvider } from '@routers/OrgRouter/context/GlobalContext'
-import WithErrorSearchContext from '@routers/OrgRouter/WithErrorSearchContext'
-import WithSessionSearchContext from '@routers/OrgRouter/WithSessionSearchContext'
+import { GlobalContextProvider } from '@routers/ProjectRouter/context/GlobalContext'
+import WithErrorSearchContext from '@routers/ProjectRouter/WithErrorSearchContext'
+import WithSessionSearchContext from '@routers/ProjectRouter/WithSessionSearchContext'
 import { auth } from '@util/auth'
 import { useIntegrated } from '@util/integrated'
 import { isOnPrem } from '@util/onPrem/onPremUtils'
