+++ frontend/src/routers/ProjectRouter/WorkspaceRedirectionRouter.tsx
